Refactor and reorganize translator codebase
Removed unused backup.py and streamlined translation logic into translate.py for a cleaner architecture. Introduced config files like .flake8 and pytest.ini to enforce conventions and testing setup. Simplified parser implementation in json.py by removing unused blocks and ensuring maintainability.
