+++ pytest.ini
@@ -1,0 +1,6 @@
+# pytest.ini
+[pytest]
+minversion = 6.0
+addopts = -ra -q
+testpaths =
+   tests
